+++ src/htsjdk/variant/bcf2/BCF2Decoder.java
@@ -282,7 +282,7 @@ public final class BCF2Decoder {
     }
 
     private double rawFloatToFloat(final int rawFloat) {
-        return (double)Float.intBitsToFloat(rawFloat);
+        return Float.intBitsToFloat(rawFloat);
     }
 
     // ----------------------------------------------------------------------
